test: stabilize mutation quick prerequisites
