+++ tests/a11y/components.vitest.test.ts
@@ -1,0 +1,6 @@
+import { describe, it, expect, beforeEach } from 'vitest';
+import { registerAccessibilityTests } from './accessibility-suite.js';
+
+describe('Accessibility (Vitest)', () => {
+  registerAccessibilityTests({ describe, it, expect, beforeEach });
+});
